Fix deleting clusters if stack is deleted
If a stack has been deleted (either by Magnum or the user) but
Magnum did not update and set `stack_id` to empty, the cluster
will fail to delete inside pre-deletion.

This will have a safe failover to skip if it can't find the
Heat stack, it assumes things are gone.

Change-Id: I6ebe188895e51ed83ad1514a380e4772fed5eb42